fix redux and make all app with redux states

+++ src/main/java/com/gate/gatelib/repository/ProblemSetDao.java
@@ -3,6 +3,6 @@ package com.gate.gatelib.repository;
 import com.gate.gatelib.models.ProblemSet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProblemSetDao extends JpaRepository<ProblemSet, Integer> {
+public interface ProblemSetDao extends JpaRepository<ProblemSet, Long> {
 
 }
